Refactoring the app
