+++ .github/workflows/azure-static-app.yml
@@ -11,6 +11,9 @@ on:
       - feature/UI-Layout
       - master
 
+env:
+  NODE_OPTIONS: "--max_old_space_size=4096"
+
 jobs:
   build_and_deploy_job:
     if: github.event_name == 'push' || (github.event_name == 'pull_request' && github.event.action != 'closed')
