+++ src/theme/overrides/Button.ts
@@ -15,9 +15,6 @@ export default function Button(theme: Theme) {
         containedInherit: {
           color: theme.palette.grey[800],
           boxShadow: theme.customShadows.z8
-          //   "&:hover": {
-          //     backgroundColor: theme.palette.grey[400]
-          //   }
         },
         containedPrimary: {
           boxShadow: theme.customShadows.primary
